Add Pawn.GetArmedWeaponType
Returns the weapon type of the armed weapon, or nil if none is armed.

The mod loader has a function Pawn.GetArmedWeapon that does the same, but it relies on reading save data, which is not always up to date.
Move Pawn.GetArmedWeapon to Pawn.GetArmedWeaponVanilla; and make Pawn:GetArmedWeapon call Pawn:GetArmedWeaponType if memedit is available.